src/main/java/com/poiasd/restphonebooks/exception/GlobalControllerExceptionHandler.java
package com.poiasd.restphonebooks.exception;

import org.springframework.http.HttpHeaders;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.MethodArgumentNotValidException;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.context.request.WebRequest;
import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;

/**
 * The global controller exception handler.
 */
@ControllerAdvice
public class GlobalControllerExceptionHandler extends ResponseEntityExceptionHandler {

    @ExceptionHandler(EntityRuntimeExceptions.EntityNotFoundException.class)
    public final ResponseEntity<Object> handleNotFoundExceptions(EntityRuntimeExceptions.EntityNotFoundException ex) {
        String message = ex.getMessage();
        return new ResponseEntity<>(message, HttpStatus.NOT_FOUND);
    }

    @ExceptionHandler(EntityRuntimeExceptions.DuplicateEntityException.class)
    public final ResponseEntity<Object> handleDuplicateExceptions(EntityRuntimeExceptions.DuplicateEntityException ex) {
        String message = ex.getMessage();
        return new ResponseEntity<>(message, HttpStatus.CONFLICT);
    }

    @Override
    protected ResponseEntity<Object> handleMethodArgumentNotValid(MethodArgumentNotValidException ex,
                                                                  HttpHeaders headers, HttpStatus status, WebRequest request) {
        StringBuilder sb = new StringBuilder();
        sb.append("Bad request - field(s):");
        ex.getFieldErrors().forEach(fieldError -> sb
                .append("\n")
                .append(fieldError.getField())
                .append(": ")
                .append(fieldError.getDefaultMessage()));
        return new ResponseEntity<>(sb.toString(), HttpStatus.BAD_REQUEST);
    }

    @ExceptionHandler(Exception.class)
    public ResponseEntity<Object> handleAllUncaughtException() {
        String message = "Unknown error occurred";
        return new ResponseEntity<>(message, HttpStatus.INTERNAL_SERVER_ERROR);
    }
}
